requirements.txt
flask==3.0.3
flask-login==0.6.3
flask-wtf==1.2.1
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
psycopg[binary]==3.3.2
python-dotenv==1.0.1
werkzeug==3.0.3
redis==5.0.8
celery==5.4.0
email-validator==2.1.2
flask-mail==0.10.0